+++ src/main/kotlin/indent/rainbow/settings/IrConfigurable.kt
@@ -35,25 +35,25 @@ class IrConfigurable : BoundConfigurable("Indent Rainbow") {
                 radioButton("Pastel (6 colors)", IrColorsPaletteType.PASTEL)
             }
             row {
-                radioButton("Spectrum (13 colors)", IrColorsPaletteType.SPECTRUM)
+                radioButton("Spectrum (6 colors)", IrColorsPaletteType.SPECTRUM)
             }
             row {
-                radioButton("Nightfall (10 colors)", IrColorsPaletteType.NIGHTFALL)
+                radioButton("Rainbow Spectrum (6 colors)", IrColorsPaletteType.RAINBOW_SPECTRUM)
+            }
+            row {
+                radioButton("Nightfall (6 colors)", IrColorsPaletteType.NIGHTFALL)
             }
             row {
                 radioButton("Aquaflow (6 colors)", IrColorsPaletteType.AQUAFLOW)
             }
             row {
-                radioButton("Luminarium (16 colors)", IrColorsPaletteType.LUMINARIUM)
+                radioButton("Luminarium (6 colors)", IrColorsPaletteType.LUMINARIUM)
             }
             row {
-                radioButton("Monochrome (9 colors)", IrColorsPaletteType.MONOCHROME)
+                radioButton("Monochrome (6 colors)", IrColorsPaletteType.MONOCHROME)
             }
             row {
-                radioButton("Prism (18 colors)", IrColorsPaletteType.PRISM)
-            }
-            row {
-                radioButton("Solarized (8 colors)", IrColorsPaletteType.SOLARIZED)
+                radioButton("Solarized (6 colors)", IrColorsPaletteType.SOLARIZED)
             }
             row {
                 val radioButton = radioButton("Custom with colors:", IrColorsPaletteType.CUSTOM)